Who is Soundboy
WHO IS SOUNDBOY?
Soundboy is anyone or anything that breaks your signal flow. Soundboy
dwells within the wrong adapter for your device, that traffic light
that is taking too long or that person that stands between you and your
immediate or long term goal.  Soundboy is the ghost in the machine.  In
Raz Mesinai’s “Soundboy Manifesto” we follow a young Producer’s
evolution within the world of sound and soundsystems.  This journey
involves trials and tribulations faced by the modern DJ / Producer and
also documents the presence of and battles with Soundboy along the way.

STORY WORLD EXCERPTS:
Selector: “Selector” is a story that follows the adventures of
waveforms and samples as they hope to stay on the grid long enough to
be selected by Producer and routed to freedom through the Master
Output. Join bassline4.wav and lilsnare.mp3 as they do what needs to be
done to survive inside the system.

Engineer: The story of the “The Engineer” involves traveling through
time to gather samples needed for the Ghost Producer’s battle against
Soundboy.  In this excerpt we go back to The Hannibal Wars where the
march of elephants is sampled for kick drums and the clash of swords is
archived to create hi hat sounds.

Ghost Producer: In “Ghost Producer” the weapon for Soundclash against
Soundboy is manifested and leads to the end of civilization as we know
it.  The Ghost Producer has been warning those who preside over the
P.A. (Producer’s Alliance) against Soundboy’s plot to suppress
frequencies and pacify the public.  As Soundboy completes his
integration of Pur-Ton, Ghost Producer plots to counter-strike with his
custom soundsystem called “The War Cry”.

THOE: In “The Heretic Of Ether” (THOE) a population has risen from the
ashes of the final Soundclash and it is up to Septa 6, the plumber for
The Main Stream, to help decide the fate of the new  mankind.  In the
world of THOE, information is liquified as opposed to stored in the
cloud.  This creates an industry for the “juice” throughout the
Quadrants.  In the underground, Rogue Blenders create “blends” from the
loose juice brought in by scavengers.  Septa 6 works to keep the pipes
in tact, but between the Scavengers, Subs and folks poking around from
the upper quadrants, Septa 6 can always count on adventures while on
duty.